# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RecoEngine API",
    "description": "RecoEngine turns any catalogue (jobs, dishes, products, courses, or your own items) into\nrecommendations. Upload items, and they are embedded with OpenAI and stored in a vector\nindex; then ask for items similar to a text, an existing item, or a profile.\n\n**Authentication:** send your API key in the `X-API-Key` header. Keys are created at\nregistration and on the API Keys page of the dashboard.\n\n**Errors** always look like `{\"error\": {\"code\": \"...\", \"message\": \"...\", \"details\": [...]}}`.\n`429` and `503` responses include a `Retry-After` header (seconds).\n\n**Rate limits:** 100 requests per minute per API key, 10,000 uploaded items per tenant per\nUTC day.",
    "version": "0.1.0",
    "contact": {
      "name": "RecoEngine",
      "url": "https://recoengine.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.recoengine.io",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/items/upload": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Upload items (JSON)",
        "operationId": "uploadItems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUploadRequest"
              },
              "examples": {
                "hr": {
                  "summary": "HR: job postings",
                  "value": {
                    "async": true,
                    "items": [
                      {
                        "external_id": "job-101",
                        "title": "Senior Python Developer",
                        "description": "Build FastAPI microservices for a hiring platform",
                        "skills": [
                          "Python",
                          "FastAPI",
                          "PostgreSQL"
                        ],
                        "location": "Bangalore",
                        "department": "Engineering",
                        "employment_type": "full_time"
                      },
                      {
                        "external_id": "job-102",
                        "title": "Data Scientist",
                        "description": "Train ranking models on user behaviour data",
                        "skills": [
                          "Python",
                          "PyTorch",
                          "SQL"
                        ],
                        "location": "Remote",
                        "department": "Data",
                        "employment_type": "full_time"
                      }
                    ]
                  }
                },
                "food": {
                  "summary": "Food: dishes, embedded before responding",
                  "value": {
                    "async": false,
                    "items": [
                      {
                        "external_id": "dish-1",
                        "name": "Penne Arrabbiata",
                        "description": "Penne in a fiery tomato, garlic and red chilli sauce",
                        "cuisine": "Italian",
                        "ingredients": [
                          "penne",
                          "tomato",
                          "chilli"
                        ],
                        "dietary_tags": [
                          "vegetarian"
                        ],
                        "price_range": "$$"
                      },
                      {
                        "external_id": "dish-2",
                        "name": "Paneer Tikka",
                        "description": "Smoky grilled cottage cheese marinated in spiced yogurt",
                        "cuisine": "North Indian",
                        "ingredients": [
                          "paneer",
                          "yogurt"
                        ],
                        "dietary_tags": [
                          "vegetarian",
                          "gluten_free"
                        ],
                        "price_range": "$$"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SyncUploadResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AsyncUploadResponse"
                    }
                  ],
                  "title": "Response Upload Items Api V1 Items Upload Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "202": {
            "description": "Queued (async=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncUploadResponse"
                },
                "example": {
                  "batch_id": "48de4e95-c0e1-4d23-b77c-a12ebfe51b42",
                  "status": "PENDING",
                  "total_items": 2,
                  "processed_items": 0,
                  "failed_items": 0,
                  "progress_percentage": 0.0,
                  "created_at": "2026-09-24T04:16:16Z",
                  "completed_at": null,
                  "mode": "async",
                  "status_url": "/api/v1/items/batch/48de4e95-c0e1-4d23-b77c-a12ebfe51b42"
                }
              }
            }
          },
          "503": {
            "description": "Embedding service down; items kept PENDING",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Upload up to 1000 items. Each needs an `external_id`; any other fields are stored, and the domain config decides which are embedded and which become filters. Re-uploading an `external_id` updates the item.\n\nWith `async: true` (default) the response is `202` with a `batch_id` to poll. With `async: false`, up to 50 items are embedded before responding."
      }
    },
    "/api/v1/items/upload-csv": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Upload items from a CSV file (always async)",
        "operationId": "uploadItemsCsv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_items_csv_api_v1_items_upload_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvUploadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid CSV, with column hints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Upload a UTF-8 CSV (up to 10,000 rows, 10 MB). Headers are matched to your fields ignoring case and punctuation; the id column may be `external_id`, `id`, `item_id` or `sku`. Always asynchronous."
      }
    },
    "/api/v1/items/batch/{batch_id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Batch progress",
        "operationId": "getBatchStatus",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Progress of an upload or rebuild. Poll until `status` is `DONE` or `PARTIAL_FAIL`."
      }
    },
    "/api/v1/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "List items (20 per page)",
        "operationId": "listItems",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EmbeddingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "description": "Part of an external_id, any case",
              "title": "Search"
            },
            "description": "Part of an external_id, any case"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Items, newest first, 20 per page. Filter by `status` or by part of the `external_id`."
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Delete ALL of this tenant's items from the database and Pinecone",
        "operationId": "deleteAllItems",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Pinecone unavailable; nothing deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Delete every item of the tenant from the database and the vector index."
      }
    },
    "/api/v1/items/bulk-delete": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Delete up to 1000 items from the database and Pinecone",
        "operationId": "bulkDeleteItems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Pinecone unavailable; nothing deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Delete up to 1000 items by `external_id`. Unknown ids are reported, not an error."
      }
    },
...
    "/api/v1/index/stats": {
      "get": {
        "tags": [
          "Index"
        ],
        "summary": "Pinecone index stats for this tenant",
        "operationId": "getIndexStats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexStatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Pinecone unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Vector count and dimension of the tenant's index, with item counts by status."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_items_csv_api_v1_items_upload_csv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "UTF-8 CSV with a header row"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_items_csv_api_v1_items_upload_csv_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ItemUploadRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "async": {
            "type": "boolean",
            "title": "Async",
            "description": "true: queue a batch and return its id. false: process up to 50 items before responding.",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ItemUploadRequest"
      },
...
    },
    "securitySchemes": {
      "AdminKeyAuth": {
        "type": "apiKey",
        "description": "Operator key (ADMIN_API_KEY). Never give it to tenants.",
        "in": "header",
        "name": "X-Admin-Key"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "Your API key, e.g. `reco_…`. Create keys in the dashboard or via /me/api-keys.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}